Fix argument parsing bug
The --clear-cache option requires an argument: YES or NO.
If shift is called only once, then the value of --clear-cache
is parsed as an argument in its own right.

This leads to either YES or NO being stored and then restored
as a positional argument. It doesn't currently seem to be harmful
but it is inaccurate and could lead to bugs in the future.