add unzip to list of Ubuntu/Debian dependencies
The Rust 1.61 version of install-rust-toolchain.sh calls unzip, which isn't installed on Ubuntu/Debian by default (at least not on our ubuntu:20.04-based Docker container).

This branch adds the utility to the list of dependencies in README.md, so users know to install it before running install-rust-toolchain.sh on Ubuntu/Debian.